webhook/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
comm==0.2.2
db-dtypes==1.3.0
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
docstring_parser==0.16
executing==2.0.1
Flask==3.0.2
frozenlist==1.4.1
functions-framework==3.8.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-aiplatform==1.46.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-dialogflow==2.30.0
google-cloud-dialogflow-cx==1.33.0
google-cloud-discoveryengine==0.12.3
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.3
parso==0.8.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyngrok==7.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.2
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.1
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.12.1
